now we have user's notifications btn, currency btn, centered btns on header.
